requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
requests==2.31.0
validators==0.22.0
python-slugify==8.0.1
Flask-User==1.0.2.2